Added a prompt so the user can chose whether or not to back up old dotfiles.

+++ deploy.sh
@@ -1,6 +1,16 @@
-mv ~/.zshrc ~/.zshrc.old
-mv ~/.tmux.conf ~/.tmux.conf.old
-mv ~/.vimrc ~/.vimrc.old
+
+echo -n "Would you like to backup your current dotfiles? (y/n) "
+old_stty_cfg=$(stty -g)
+stty raw -echo
+answer=$( while ! head -c 1 | grep -i '[ny]' ;do true ;done )
+stty $old_stty_cfg
+if echo "$answer" | grep -iq "^y" ;then
+	mv ~/.zshrc ~/.zshrc.old
+	mv ~/.tmux.conf ~/.tmux.conf.old
+	mv ~/.vimrc ~/.vimrc.old
+else
+	echo -e "\nNot backing up old dotfiles."
+fi
 
 printf "source '$HOME/dotfiles/zsh/zshrc_manager.sh'" > ~/.zshrc
 printf "so $HOME/dotfiles/vim/vimrc.vim" > ~/.vimrc
